test: ImageSize 테스트코드
