+++ src/test/java/nextstep/session/image/ImageCapacityTest.java
@@ -1,0 +1,18 @@
+package nextstep.session.image;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+class ImageCapacityTest {
+
+    @DisplayName("이미지가 1MB를 초과하면 예외가 발생한다.")
+    @Test
+    void createImageOverCapacityThrowExceptionTest() {
+
+        assertThatThrownBy(() -> new ImageCapacity(2))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("이미지 사이즈는 1MB를 초과하면 안됩니다.");
+    }
+}
