+++ src/test/java/nextstep/session/image/ImageSizeTest.java
@@ -7,12 +7,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ImageSizeTest {
 
-    @DisplayName("이미지가 1MB를 초과하면 예외가 발생한다.")
+    @DisplayName("이미지의 너비 높이가 3:2 비율이 아니라면 예외가 발생한다.")
     @Test
-    void createImageOverSizeThrowExceptionTest() {
+    void createImageRateThrowExceptionTest() {
 
-        assertThatThrownBy(() -> new ImageSize(2))
+        assertThatThrownBy(() -> new ImageSize(200, 200))
                 .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이미지 사이즈는 1MB를 초과하면 안됩니다.");
+                .hasMessage("이미지의 너비 높이가 3:2 비율이여야 합니다.");
     }
 }
